+++ ddsrouter_core/test/unittest/efficiency/MapPayloadPoolTest.cpp
@@ -25,7 +25,7 @@ using namespace eprosima::ddsrouter;
 using namespace eprosima::ddsrouter::core;
 using namespace eprosima::ddsrouter::core::types;
 
-const constexpr uint16_t TEST_NUMBER = 5;
+const constexpr unsigned int TEST_NUMBER = 5;
 const constexpr size_t DEFAULT_SIZE = sizeof(PayloadUnit);
 
 namespace eprosima {
@@ -88,13 +88,13 @@ TEST(MapPayloadPoolTest, get_payload)
         test::MockMapPayloadPool pool;
         std::vector<Payload> payloads(TEST_NUMBER);
 
-        for (int i = 0; i < TEST_NUMBER; i++)
+        for (unsigned int i = 0; i < TEST_NUMBER; i++)
         {
             pool.get_payload(DEFAULT_SIZE, payloads[i]);
 
             ASSERT_EQ(payloads[i].max_size, DEFAULT_SIZE);
             ASSERT_EQ(pool.pointers_stored(), i + 1);
-            ASSERT_EQ(pool.reference_count(payloads[i]), 1);
+            ASSERT_EQ(pool.reference_count(payloads[i]), 1u);
         }
 
         // END : Clean all remaining payloads
@@ -137,46 +137,46 @@ TEST(MapPayloadPoolTest, get_payload_from_src)
 
     // get payload0
     ASSERT_TRUE(pool_->get_payload(DEFAULT_SIZE, payload0));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->reference_count(payload0), 1);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->reference_count(payload0), 1u);
 
     // get payload1 from src payload0
     ASSERT_TRUE(pool_->get_payload(payload0, pool, payload1));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->reference_count(payload1), 2);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->reference_count(payload1), 2u);
     ASSERT_EQ(payload1.max_size, payload0.max_size);
     ASSERT_EQ(payload1.data, payload0.data);
 
     // get payload2 from src payload1
     ASSERT_TRUE(pool_->get_payload(payload1, pool, payload2));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->reference_count(payload2), 3);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->reference_count(payload2), 3u);
     ASSERT_EQ(payload2.max_size, payload0.max_size);
     ASSERT_EQ(payload2.data, payload0.data);
 
     // release payload0
     ASSERT_TRUE(pool_->release_payload(payload0));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->reference_count(payload2), 2);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->reference_count(payload2), 2u);
 
     // get payload3 from src payload1
     ASSERT_TRUE(pool_->get_payload(payload1, pool, payload3));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->reference_count(payload3), 3);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->reference_count(payload3), 3u);
     ASSERT_EQ(payload3.max_size, payload1.max_size);
     ASSERT_EQ(payload3.data, payload1.data);
 
     // get payload4
     ASSERT_TRUE(pool_->get_payload(DEFAULT_SIZE * 0x100, payload4));
-    ASSERT_EQ(pool_->pointers_stored(), 2);
-    ASSERT_EQ(pool_->reference_count(payload1), 3);
-    ASSERT_EQ(pool_->reference_count(payload4), 1);
+    ASSERT_EQ(pool_->pointers_stored(), 2u);
+    ASSERT_EQ(pool_->reference_count(payload1), 3u);
+    ASSERT_EQ(pool_->reference_count(payload4), 1u);
 
     // get payload5 from src payload4
     ASSERT_TRUE(pool_->get_payload(payload4, pool, payload5));
-    ASSERT_EQ(pool_->pointers_stored(), 2);
-    ASSERT_EQ(pool_->reference_count(payload1), 3);
-    ASSERT_EQ(pool_->reference_count(payload5), 2);
+    ASSERT_EQ(pool_->pointers_stored(), 2u);
+    ASSERT_EQ(pool_->reference_count(payload1), 3u);
+    ASSERT_EQ(pool_->reference_count(payload5), 2u);
     ASSERT_EQ(payload5.max_size, payload4.max_size);
     ASSERT_EQ(payload5.data, payload4.data);
 
@@ -189,7 +189,7 @@ TEST(MapPayloadPoolTest, get_payload_from_src)
 
     // Check payload pool is empty
     ASSERT_TRUE(pool_->is_clean());
-    ASSERT_EQ(pool_->pointers_stored(), 0);
+    ASSERT_EQ(pool_->pointers_stored(), 0u);
 
     delete pool;
 }
@@ -217,21 +217,21 @@ TEST(MapPayloadPoolTest, get_payload_from_src_no_owner)
 
     // get payload aux from pool aux
     pool_aux_->get_payload(DEFAULT_SIZE, payload_src);
-    ASSERT_EQ(pool_aux_->pointers_stored(), 1);
-    ASSERT_EQ(pool_->pointers_stored(), 0);
+    ASSERT_EQ(pool_aux_->pointers_stored(), 1u);
+    ASSERT_EQ(pool_->pointers_stored(), 0u);
 
     // get payload from src payload aux
     ASSERT_TRUE(pool_->get_payload(payload_src, pool_aux, payload_target));
-    ASSERT_EQ(pool_->pointers_stored(), 1);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
 
     // release payload aux from pool aux
     pool_aux_->release_payload(payload_src);
-    ASSERT_EQ(pool_aux_->pointers_stored(), 0);
-    ASSERT_EQ(pool_->pointers_stored(), 1);
+    ASSERT_EQ(pool_aux_->pointers_stored(), 0u);
+    ASSERT_EQ(pool_->pointers_stored(), 1u);
 
     // release payload
     pool_->release_payload(payload_target);
-    ASSERT_EQ(pool_->pointers_stored(), 0);
+    ASSERT_EQ(pool_->pointers_stored(), 0u);
 
     delete pool_aux;
     delete pool;
@@ -306,39 +306,39 @@ TEST(MapPayloadPoolTest, release_payload)
     pool_->get_payload(DEFAULT_SIZE, payloads[0]);
 
     // get N-1 payloads from first
-    for (int i = 1; i < TEST_NUMBER; i++)
+    for (unsigned int i = 1; i < TEST_NUMBER; i++)
     {
         pool_->get_payload(payloads[0], pool, payloads[i]);
         ASSERT_EQ(pool_->reference_count(payloads[0]), i + 1) << i;
     }
 
     // release N-2 payloads
-    for (int i = 2; i < TEST_NUMBER; i++)
+    for (unsigned int i = 2; i < TEST_NUMBER; i++)
     {
         ASSERT_TRUE(pool_->release_payload(payloads[i]));
         ASSERT_EQ(pool_->reference_count(payloads[0]), TEST_NUMBER + 1 - i) << i;
     }
 
     // get N-2 more payloads from first
-    for (int i = 2; i < TEST_NUMBER; i++)
+    for (unsigned int i = 2; i < TEST_NUMBER; i++)
     {
         pool_->get_payload(payloads[0], pool, payloads[i]);
         ASSERT_EQ(pool_->reference_count(payloads[0]), i + 1) << i;
     }
 
     // release N payloads
-    for (int i = 1; i < TEST_NUMBER; i++)
+    for (unsigned int i = 1; i < TEST_NUMBER; i++)
     {
         ASSERT_TRUE(pool_->release_payload(payloads[i]));
         ASSERT_EQ(pool_->reference_count(payloads[0]), TEST_NUMBER - i) << i;
     }
     // Removing last payload because if not the reference count cannot be done
     ASSERT_TRUE(pool_->release_payload(payloads[0]));
-    ASSERT_EQ(pool_->pointers_stored(), 0);
+    ASSERT_EQ(pool_->pointers_stored(), 0u);
 
     // Check payload pool is empty
     ASSERT_TRUE(pool_->is_clean());
-    ASSERT_EQ(pool_->pointers_stored(), 0);
+    ASSERT_EQ(pool_->pointers_stored(), 0u);
 
     delete pool;
 }
